feat: MemoFlow 문서 저장, 비동기 AI 처리 및 상태 기반 UI 구조 구현

+++ app/src/main/java/com/example/memoflow/ui/navigation/Destinations.kt
@@ -6,6 +6,6 @@ sealed class Destinations(val route: String) {
     data object Search: Destinations("search")
     data object Settings: Destinations("settings")
     data object Detail: Destinations("detail/{documentId}"){
-        fun createRoute(documentId: Int) = "detail/$documentId"
+        fun createRoute(documentId: Long) = "detail/$documentId"
     }
 }
